исправление всех ошибок после проверки проекта наставником
