+++ js/add-thumbnails.js
@@ -1,12 +1,15 @@
-import {picturesContainerElement, gallareyTemplate} from './project-wide-search-elements.js';
+const picturesContainerElement = document.querySelector('.pictures');
+const gallareyTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
 const addThumbnails = (({id, url, description, likes, comments}) => {
   const thumbnail = gallareyTemplate.cloneNode(true);
+
   thumbnail.dataset.pictureId = id;
   thumbnail.querySelector('.picture__img').src = url;
   thumbnail.querySelector('.picture__img').alt = description;
   thumbnail.querySelector('.picture__likes').textContent = likes;
   thumbnail.querySelector('.picture__comments').textContent = comments.length;
+
   picturesContainerElement.appendChild(thumbnail);
 });
